Encrypted database table view.

+++ pdm/pdm_database.h
@@ -21,7 +21,7 @@ public:
   pdm_database();
   ~pdm_database();
 
-  int open_db(char* name, char*pas,size_t pas_size);
+  int open_db(char* name, char*pas,int pas_size);
   int close_db(char* name);
   int execute(char *input);
